queuer/requirements.txt
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
idna==3.10
isodate==0.7.2
lxml==6.0.2
more-itertools==10.8.0
nexus-rpc==1.1.0
platformdirs==4.4.0
protobuf==5.29.5
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
simple-salesforce==1.12.9
temporalio==1.18.0
trino==0.336.0
types-protobuf==6.32.1.20250918
typing_extensions==4.15.0
urllib3==2.5.0
zeep==4.3.2
